+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cobaltcore-dev/dashboard
 
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [6dd15fd]
+  - @cobaltcore-dev/aurora@0.20.2
+
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [7549dec]
+- Updated dependencies [04cc26d]
+  - @cobaltcore-dev/aurora@0.20.1
+
 ## 1.2.2
 
 ### Patch Changes